requirements.txt
aiogram==3.13.1
APScheduler==3.10.4
SQLAlchemy==2.0.36
python-dotenv==1.0.1
google-generativeai==0.8.3
pytz==2024.2
